app/requirements.txt
google-genai
xmltodict
lxml
requests
beautifulsoup4
fastapi
uvicorn
rapidfuzz
python-dotenv
elasticsearch